ai_agents/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
gunicorn
langchain
langgraph
langchain-groq
pymongo[srv]
redis
# For vector search with Redis
langchain-redis
# For embedding text
langchain-openai